+++ Backend/src/main/java/com/Marketplace_Management/Catalog/DTOs/Requests/Product/CreateProductVariantRequest.java
@@ -34,5 +34,5 @@ public class CreateProductVariantRequest {
     private List<MultipartFile> images;
 
     @Distinct
-    private List<Integer> optionIds;
+    private List<Long> optionIds;
 }
